Verschlanken der boot.php, Code ausgelagert
